+++ lib/src/dartbuilders/mockor_dart_builder.dart
@@ -1,6 +1,6 @@
 import 'package:code_builder/code_builder.dart';
 import 'package:dart_style/dart_style.dart';
-import 'package:mockor/annotations.dart';
+import 'package:mockor/mockor.dart';
 import 'package:mockor/src/models/models.dart';
 
 ///Builds Dart code of the mocker function