+++ lib/src/generators/common/common.dart
@@ -1,7 +1,7 @@
 import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
-import 'package:mockor/annotations.dart';
+import 'package:mockor/mockor.dart';
 import 'package:mockor/src/exceptions/mockor_exception.dart';
 import 'package:mockor/src/models/models.dart';
 import 'package:source_gen/source_gen.dart';