+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improve package description
 - Move builder.dart to src folder.
 - Fix compatibility with analyzer 1.7.0
+- Change annotations package to mockor.
 
 ## 1.0.0